Move M-H test, clarify comments
